Clarify that there are two distinct types of cf object
Fixes #15.

There are two distinct cf objects serves associated with a Request:
1. On an eyeball Request attached to a FetchEvent, the cf property
   contains metadata about the request, provided by Cloudflare's
   edge.
2. As part of a dictionary of configuration passed to the Request
   constructor, the cf property allows a user to control certain
   Cloudflare functionality that can be applied to said Request,
   like image resizing.

While those two things are both objects in a dictionary with the
key "cf", they are otherwise distinct.

This change makes these two types explicitly different, and
specifies that a Request constructor's RequestInit dict can have
a "cf" property that is either (1) or (2).

Previously, passing a Request object as the `init` arg to the
Request constructor failed, because an eyeball request's cf prop
did not have the same shape as (2).