Add encodeBody property to Response
The [Cloudflare specific property `encodeBody`][0] is missing from both
`Response` and `ResponseInit` types.  This commit fixes this by adding
the property to the types. I currently use the `ResponseInit` option in
production so I know that works, and I just spun up a test worker for
`Response#encodeBody` and saw that when unassigned, it returns null.

[0]: https://developers.cloudflare.com/workers/runtime-apis/response#properties