Move `cf` types to `manual-ts` (#306)
The current version of `workers-types` doesn't support `&` in overrides :cry:,
so this PR moves these types to the `manual-ts` folder, which gets
concatenated onto the generated types. Given none of these types are defined
in the runtime, we don't need to override them, so this will work for now.

The new version of `workers-types` doesn't have this restriction, so this won't
be a problem soon.